Refactor Craftscript loading
