+++ src/main/java/fourteener/worldeditor/worldeditor/macros/macros/nature/ErodeMacro.java
@@ -1,4 +1,4 @@
-package fourteener.worldeditor.worldeditor.macros.macros;
+package fourteener.worldeditor.worldeditor.macros.macros.nature;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,6 +12,7 @@ import org.bukkit.block.BlockState;
 
 import fourteener.worldeditor.main.Main;
 import fourteener.worldeditor.operations.Operator;
+import fourteener.worldeditor.worldeditor.macros.macros.Macro;
 import fourteener.worldeditor.worldeditor.undo.UndoElement;
 import fourteener.worldeditor.worldeditor.undo.UndoManager;
